+++ src/java/facade/IPersonFacade.java
@@ -14,7 +14,7 @@ import java.util.List;
  */
 public interface IPersonFacade { 
     public Person addPerson(Person p); 
-    public Person deletePerson(int id); 
+    public Person deletePerson(long id); 
     public Person getPerson(long id); 
     public List<Person> getPersons(); 
     public Person editPerson(Person p);
